Basic C2000 build
This hacks out any non-cross -platform code inside
openinverter and makes the code build. It's non functional
and there are many warnings but it links for FLASH. RAM
builds fail at the link stage due to lack of space (no
optimisation also).

+++ libopeninv/include/stm32_can.h
@@ -68,7 +68,7 @@ private:
       uint16_t mapParam;
       s16fp gain;
       uint8_t offsetBits;
-      int8_t numBits;
+      int16_t numBits;
    };
 
    struct CANIDMAP
